+++ examples/resources/pingfederate_keypairs_signing_csr_response/resource.tf
@@ -1,0 +1,4 @@
+resource "pingfederate_keypairs_signing_csr_response" "csrResponse" {
+  keypair_id = "mysigningkeypair"
+  file_data  = filebase64("./path/to/csr_response.pem")
+}