+++ scripts/verifyContent.py
@@ -9,6 +9,7 @@ missing_changelog_entries = set()
 
 noImportResources = [
   "pingfederate_connection_metadata_export",
+  "pingfederate_keypairs_signing_csr_response",
   "pingfederate_keypairs_signing_key",
   "pingfederate_keypairs_ssl_client_csr_response",
   "pingfederate_keypairs_ssl_client_csr_export",